Reorganized photo directory
Moved all photos belonging to landing page to its own folder
