+++ front-end/public/images/landingPage/aboutUs/icon_github.png
