+++ front-end/public/images/landingPage/features/feature2.png
