+++ front-end/public/images/landingPage/features/feature3.png
