+++ front-end/public/images/landingPage/hero/hero.png
